\name{pbdDEMO-package}
\alias{pbdDEMO-package}
\docType{package}
\title{Demonstrations and Examples for the pbd Project}
\description{
  Demos
}
\details{
\tabular{ll}{
  Package: \tab pbdDMAC\cr
  Type: \tab Package\cr
  License: \tab GPL\cr
  LazyLoad: \tab yes\cr
}
  This package requires an MPI library (OpenMPI, MPICH2, or LAM/MPI).
}
\references{
  Programming with Big Data in R Website:
  \url{http://r-pbd.org/}
}
\author{
  Drew Schmidt \email{schmidt AT math.utk.edu}, Wei-Chen Chen,
  George Ostrouchov, and Pragneshkumar Patel.
}
\keyword{Package}
