% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/control.r
\docType{data}
\name{DEMO Control}
\alias{.DEMO.CT}
\alias{DEMO Control}
\title{A set of controls in pbdDEMO}
\format{Objects contain several default parameters for BLACS.}
\usage{
.DEMO.CT
}
\description{
This set of controls is used to provide default values in this package.
}
\details{
The elements of \code{.DEMO.CT} are default values \tabular{lcl}{ Elements
\tab Default \tab Usage \cr \code{gbd.major} \tab 1L \tab a default GBD
row-major \cr \code{ictxt} \tab 0L \tab a default BLACS context \cr
\code{bldim} \tab c(2L,2L) \tab a default block dimension \cr
\code{divide.method} \tab "block.cyclic" \tab a default balance method \cr }
}
\keyword{global}
\keyword{variables}

