% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pbdDEMO-package.r
\docType{package}
\name{pbdDEMO-package}
\alias{pbdDEMO-package}
\title{Demonstrations and Examples for the pbd Project}
\description{
A set of demos of pbdR packages, together with a useful,
unifying vignette.
}
\details{
\tabular{ll}{ Package: \tab pbdDEMO\cr Type: \tab Package\cr License: \tab
MPL 2.0\cr LazyLoad: \tab yes\cr } This package requires an MPI library
(OpenMPI, MPICH2, or LAM/MPI).
}
\author{
Drew Schmidt \email{schmidt AT math.utk.edu}, Wei-Chen Chen, George
Ostrouchov, and Pragneshkumar Patel.
}
\references{
Programming with Big Data in R Website: \url{http://r-pbd.org/}
}
\keyword{Package}

