% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timer.R
\name{timer}
\alias{timer}
\title{A Timing Function for SPMD Routines}
\usage{
timer(timed)
}
\arguments{
\item{timed}{expression to be timed.}
}
\value{
A named vector containing the minimum, mean, and maximum time across
all processors in the communicator.  All values are global.
}
\description{
A timing function for use with parallel codes executed in the batch SPMD
style.
}
\details{
Finds the min, mean, and max execution time across all independent processes
executing the operation \code{timed}.
}
\keyword{Timing}

