\name{demo.internal}

\alias{nbd}
\alias{demo.ncvar_ndim}
\alias{demo.ncvar_put_2D}
\alias{demo.ncvar_get_2D}
\alias{demo.gbdr2dmat}
\alias{demo.gbdc2dmat}
\alias{demo.dmat2gbdr}
\alias{demo.dmat2gbdc}

\title{ All pbdDEMO Internal Functions}
\description{ All pbdDEMO internal functions }
\keyword{internal}
