\name{demo.internal}

\alias{nbd}
\alias{demo.ncvar_ndim}
\alias{demo.ncvar_put_2D}
\alias{demo.ncvar_get_2D}
\alias{demo.spmdr2dmat}
\alias{demo.spmdc2dmat}
\alias{demo.dmat2spmdr}
\alias{demo.dmat2spmdc}

\title{ All pbdDEMO Internal Functions}
\description{ All pbdDEMO internal functions }
\keyword{internal}
